test(checkpoint-postgres): remove overlapping SQL mock coverage
